Add js-beautify to HTML code blocks for component pages, clean up Command Button HBS
