+++ src/documentation/pages/Components/SearchBox/examples/SearchBoxExampleModel.js
@@ -1,0 +1,16 @@
+var SearchBoxExampleProps = {
+  "label": "Search",
+  "defaultValue": "",
+  "icon": "Search",
+  "clearButton": {
+    "component": "CommandButton",
+    "props": {
+      "icon": "Clear",
+      "modifier": "noLabel",
+      "tag": "button",
+      "customClasses": "ms-SearchBox-clear"
+    }
+  }
+}
+
+module.exports = SearchBoxExampleProps;
